Rename ext for rollup config from "js" to "mjs"
